requirements.txt
netaddr
netifaces
python-telegram-bot
python-nmap
requests
mitmproxy